% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLabel.R
\name{convertLabel}
\alias{convertLabel}
\title{Transform SPSS variable and value labels.}
\usage{
convertLabel(spssList, stringsAsFactors = TRUE, useZkdConvention = TRUE)
}
\arguments{
\item{spssList}{An object created by \code{\link[foreign]{read.spss}}. Important: Using \code{\link[foreign]{read.spss}},
\code{to.data.frame} has to be \code{FALSE}.}

\item{stringsAsFactors}{Transform character variables into factors?}

\item{useZkdConvention}{Logical: If \code{TRUE}. variable labels are named \code{'varLabel'}, value labels are named \code{'valLabel'}.}
}
\value{
A \code{data.frame}.
}
\description{
By default, \code{\link[foreign]{read.spss}} from the \code{foreign} package uses variable labels as attributes of the whole
\code{data.frame}, value labels as attribute of each specific
variable in the \code{data.frame}. \code{convertLabel} provides variable and value labels as variable attributes.
}
\examples{
file <- system.file("extdata", "Klauer.sav", package = "eatAnalysis")
dat  <- foreign::read.spss(file, to.data.frame=FALSE, use.value.labels = FALSE)
dat  <- convertLabel(dat)
str(dat)
}
\author{
Sebastian Weirich
}
