% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyScanSelection.R
\name{copyScanSelection}
\alias{copyScanSelection}
\title{Copy booklet scan selection to desired directory}
\usage{
copyScanSelection(
  vars,
  dat,
  id,
  sourceDir,
  targetDir,
  codebook,
  startRow = 4,
  sheet = "Codebook",
  varColumn = "Variable",
  bookletColumnPrefix = "TH",
  separators = c("-", "_"),
  suffix = ".TIF"
)
}
\arguments{
\item{vars}{character string of variables whose scans should be copied.}

\item{dat}{wide format data which contain these variables. Names in \code{vars} must match column names of \code{dat}.}

\item{id}{column number or column name of the person identifier variable in the wide format data.}

\item{sourceDir}{The directory which contains the scans.}

\item{targetDir}{The target directory for the copied scans.}

\item{codebook}{either a character string with the folder of the corresponding excel file,
or an already imported data.frame of the IQB codebook.}

\item{startRow}{Optional: If codebook is provided as character string referring to an excel file,
\code{startRow} indicates the first line in the file which should be read in.}

\item{sheet}{character string containing the name of the excel sheet containing the codebook.
Only necessary if \code{codebook} was provided as character string.}

\item{varColumn}{character string of the variable identifier column name in the codebook.}

\item{bookletColumnPrefix}{Character string of the booklet identifier prefix in the codebook. The codebook
usually contains several booklet columns which should begin with a common identifier.}

\item{separators}{Character strings which separates booklet identifier and person identifier in the
filenames of the scans.}

\item{suffix}{Suffix of the filenames of the scans.}
}
\value{
No return, the files will be written on disk.
}
\description{
Functions allows to copy the scans of a selection of variables,
sort by variable codes, into a desired directory.
}
\examples{
\dontrun{
# source directory
path <- "s:/Vera3-Scans/Deutsch/V3_Pilot_2015/Depot_100"
# target directory
target <- "N:/archiv/test"
# codebook folder
codebook <- "p:/R/Material/V3-2016_Codebook_Zoowaerter.xlsx"
# variable list
vars <- readxl::read_excel("p:/R/Material/KA3_Variablennamen_Zoowaerter.xlsx",
                  sheet = "Tabelle1")
vars <- substr(unique(unlist(vars)),1,7)
# load data and reshape to the wide format
load("r:/VERA3/Deutsch/V3_DEU_2016/1_Pilotierung_2015/13_Auswertung und Itemselektion/02_Itemebene.rda")
dat <- reshape2::dcast(datAggL, ID~item, value.var = "value")
# select and copy scans
cop  <- copyScanSelection(vars=vars, dat=dat, id="ID", sourceDir=path,
                         targetDir=target, codebook=codebook, startRow = 1)
}

}
\author{
Sebastian Weirich
}
