% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLabelList.R
\name{createLabelList}
\alias{createLabelList}
\title{Creates a data.frame with variable and value labels from variable attributes}
\usage{
createLabelList(dfr)
}
\arguments{
\item{dfr}{A data.frame with variable and value labels stored as attributes using the convention of \code{\link{convertLabel}}.}
}
\value{
A \code{data.frame}.
}
\description{
By default, \code{\link[foreign]{read.spss}} from the foreign package uses variable.labels
as attributes of the whole \code{data.frame}, value.labels as attribute of each specific
variable in the \code{data.frame}. \code{\link{convertLabel}} provides variable and value
labels as variable attributes. \code{createLabelList} creates a data.frame with
variable and value labels.
}
\examples{
\dontrun{
 dat <- foreign::read.spss("c:/exampleFolder/exampleDataSet.sav",
                           to.data.frame=FALSE, use.value.labels = FALSE,
                           reencode = "65001")
 dat <- convertLabel(dat)
 atts<- createLabelList(dat)
}

}
\author{
Sebastian Weirich
}
