% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLoginXml.r
\name{createLoginXml}
\alias{createLoginXml}
\title{Creates xml login files from data.frame for computer-based assessment}
\usage{
createLoginXml(
  dat,
  login,
  password,
  label,
  group = NULL,
  class = "Klasse",
  dir,
  prefix = "test",
  sep = "_",
  booklet,
  mode = "run-hot-return",
  login.mode = "monitor-group"
)
}
\arguments{
\item{dat}{data.frame with login information. See examples.}

\item{login}{Variable name or column number of the login column in \code{dat}. Login variable must not contain any missing values.}

\item{password}{Variable name or column number of the password column in \code{dat}. Passwords must not contain any missing values.}

\item{label}{Variable name or column number of the label column in \code{dat}. Missing values are not allowed.}

\item{group}{Optional: if several classes or test groups are stored in \code{dat}, \code{group} specifies variable name or column
number of the group variable in \code{dat}. If specified, missing values are not allowed.}

\item{class}{Variable name or column number of the class column in \code{dat}. Missing values are not allowed.}

\item{dir}{target directory for the xml files}

\item{prefix}{prefix for file name of the xml files}

\item{sep}{separator which separates prefix and remaining xml file name}

\item{booklet}{single string (length 1) for the desired booklet entry in the xml files. See example.}

\item{mode}{single string (length 1) for the desired mode entry in the xml files. See example.}

\item{login.mode}{single string (length 1) for the desired login mode entry in the xml files. See example.}
}
\value{
No return, the files will be written on disk.
}
\description{
In computer-based assessment, logins for each student have to be created beforehand.
If logins and password are captured in a common data.frame, function creates several xml
files, one for each class or test group.
}
\examples{
file <- system.file("extdata", "logins.xlsx", package = "eatAnalysis")
dat  <- readxl::read_excel(file, sheet = "Tabelle1")
createLoginXml (dat=dat, login = "Name", password="Passwort", label = "Label",group = "groupID",
   dir = tempdir(), prefix = "logins", sep="_",booklet = "V8DeuTBAPilot2022TH15Faultier")

}
