% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSpssSyntaxSnippet.R
\name{createSpssSyntaxSnippet}
\alias{createSpssSyntaxSnippet}
\title{Prepare SPSS syntax to transfer variable and value labels into an SPSS data set}
\usage{
createSpssSyntaxSnippet(dat, file, keep = TRUE)
}
\arguments{
\item{dat}{R data frame. Variable and value labels must be stored in attributes as provided, for example, by the function \code{\link{convertLabel}}.}

\item{file}{Character string with the name of syntax file which should be created.}

\item{keep}{Logical: Create value labels even if the value does not occur in the data? For example, if
\code{keep = TRUE}, \code{createSpssSyntaxSnippet} will write labels also for males, even if the data only contains females.}
}
\value{
No return, the SPSS syntax file is written to disk.
}
\description{
Usually, variable and value labels are lost if ASCII data is imported into SPSS.
Variable and value labels normally must be reestablished using SPSS syntax file. The function
prepares an SPSS syntax snippet from the variable and value labels stored as attributes in R.
}
\examples{
file <- system.file("extdata", "Klauer.sav", package = "eatAnalysis")
dat  <- foreign::read.spss(file, to.data.frame=FALSE, use.value.labels = FALSE)
dat  <- convertLabel(dat)
createSpssSyntaxSnippet( dat = dat, file = file.path(tempdir(), "labels.txt"), keep = FALSE)
}
\author{
Sebastian Weirich
}
