% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2tex_xtable.R
\name{df2tex_xtable}
\alias{df2tex_xtable}
\title{Create latex table code.}
\usage{
df2tex_xtable(
  df,
  caption = "",
  san = TRUE,
  label = NULL,
  note = NULL,
  note_width = 14
)
}
\arguments{
\item{df}{A \code{data.frame}.}

\item{caption}{A single character, containing the table caption.}

\item{san}{Should all characters be sanitized.}

\item{label}{A single character, containing the table label (only for referencing in the \code{.tex} document.}

\item{note}{A single character, containing the table notes.}

\item{note_width}{A single numeric, containing the table note width.}
}
\value{
Returns \code{NULL}.
}
\description{
Create latex code for a table from a \code{data.frame}.
}
\details{
Prints the resulting latex code to console
}
\examples{
# Example data set
#to be done

}
