% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.lmer.effects.R
\name{get.lmer.effects.forBootMer}
\alias{get.lmer.effects.forBootMer}
\title{\code{get.lmer.effects} for bootstrap}
\usage{
get.lmer.effects.forBootMer(lmerObj)
}
\arguments{
\item{lmerObj}{An object of class \code{merMod} or \code{glmerMod}, as created by \code{\link[lme4]{lmer}} or \code{\link[lme4]{glmer}} from the \code{lme4} package.}
}
\value{
A numerical vector of values.
}
\description{
This is a wrapper for the \code{\link{get.lmer.effects}} only intended to prepare bootstrap analyses.
See examples of \code{\link{get.lmer.effects}} for further details.
}
\examples{
### See examples of 'get.lmer.effects' for further details.

}
\author{
Sebastian Weirich
}
