% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.logit.R
\name{item.logit}
\alias{item.logit}
\title{Simulates 1PL or 2PL responses}
\usage{
item.logit(z, slope = 1, thr)
}
\arguments{
\item{z}{A vector of item difficulties.}

\item{slope}{A vector of item discrimination parameters. Default is 1 and equals a 1PL model.}

\item{thr}{A vector of item threshold parameters.}
}
\value{
A list with two matrices. First matrix includes the responses, seconds matrix includes the correct response probabilities.
}
\description{
Simulates responses following a 1PL or 2PL model given item parameters assuming normally distributed person parameters.
}
\examples{
z <- rnorm(30, 0, 1)
d <- item.logit(z = z, thr = 0)

}
\author{
Ulf Kroehne
}
