% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRdataLikeRDS.r
\name{loadRdataLikeRDS}
\alias{loadRdataLikeRDS}
\title{Reload Saved rda or Rdata Datasets and assign them to a new object}
\usage{
loadRdataLikeRDS(file)
}
\arguments{
\item{file}{the name of the rda or Rdata file}
}
\value{
If the Rdata file contains only one R object, the object is returned. If the file contains
several objects, a list of these objects is returned.
}
\description{
Function allows to load rda and Rdata objects like rds objects.
Hence, the function mimics the \code{\link[base]{readRDS}} functionality for rda and Rdata objects.
}
