% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{pretty_table}
\alias{pretty_table}
\title{Create a standard frequency table.}
\usage{
pretty_table(x, x_label, useNA = "ifany")
}
\arguments{
\item{x}{A vector.}

\item{x_label}{Label for the x variable?}

\item{useNA}{How should missing values be treated in the table? Possible values: \code{no}, \code{ifany}, \code{always}.}
}
\value{
Returns a data.frame with 2 columns, the levels of x and its frequencies.
}
\description{
Create a standard frequency table that can be written to excel.
}
\details{
Standard \code{table} output is very ugly when written to excel. This function creates an easy to write to excel object.
}
\examples{
pretty_table(iris$Species, "Flower Species")

}
