% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{prop_table}
\alias{prop_table}
\title{Create a standard relative frequency table.}
\usage{
prop_table(vec, weights = NULL, na.rm = FALSE, round_perc = 1)
}
\arguments{
\item{vec}{A vector.}

\item{weights}{A numeric vector of non-negative weights.}

\item{na.rm}{Set to \code{FALSE} to suppress checking for \code{NAs}. If \code{TRUE}, \code{NAs} are removed from \code{x} as well as from \code{weights} prior to variance estimation.}

\item{round_perc}{With how many decimals should the percentages be given?}
}
\value{
Returns a data.frame with 1 row.
}
\description{
Create a simple relative frequency table with percentages, optionally while using sampling weights.
}
\examples{
# without weights
prop_table(mtcars$cyl)

# with weights
prop_table(mtcars$cyl, weights = rep(c(0.5, 1), 16))

}
