% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{prop_table_by_and_all}
\alias{prop_table_by_and_all}
\title{Create a standard relative frequency table.}
\usage{
prop_table_by_and_all(
  df,
  dep,
  by_var,
  weights = NULL,
  na.rm = TRUE,
  round_perc = 1
)
}
\arguments{
\item{df}{A \code{data.frame}.}

\item{dep}{The name of the dependent variable in the \code{data.frame}.}

\item{by_var}{The name of the group variable in the \code{data.frame}.}

\item{weights}{The name of the weights variable in the \code{data.frame}.}

\item{na.rm}{Set to \code{FALSE} to suppress checking for \code{NAs}. If \code{TRUE}, \code{NAs} are removed from \code{x} as well as from \code{weights} prior to variance estimation.}

\item{round_perc}{With how many decimals should the percentages be given?}
}
\value{
Returns a data.frame with 1 row.
}
\description{
Create a simple relative frequency table with percentages.
}
\details{
Function is meant to be able to easily \code{rbind} various tables.
}
\examples{
# without weights
prop_table_by_and_all(df = mtcars, dep = "cyl", by_var = "gear")

# with weights
mtcars2 <- mtcars
mtcars2[, "weights"] <- rep(c(0.5, 1), 16)
prop_table_by_and_all(df = mtcars2, dep = "cyl", by_var = "gear", weights = "weights")

}
