% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xlsx.R
\name{read_xlsx}
\alias{read_xlsx}
\title{Read an \code{.xlsx} file.}
\usage{
read_xlsx(filePath)
}
\arguments{
\item{filePath}{The path to the output file.}
}
\value{
A list or a \code{data.frame}.
}
\description{
Read an \code{.xlsx} file with multiple sheets.
}
\details{
If the \code{.xlsx} contains multiple sheets, the output is a list with all sheets as
\code{data.frame} entries in the list.
}
\examples{
f <- tempfile(fileext = ".xlsx")
write_xlsx(list(cars = mtcars, flowers = iris), filePath = f)
allDat <- read_xlsx(f)

}
