% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transferAttributes.r
\name{transferAttributes}
\alias{transferAttributes}
\title{Transfer attributes from a origin data.frame to a target data.frame.}
\usage{
transferAttributes(origin, target, whichVars = NULL, whichAttrs = NULL)
}
\arguments{
\item{origin}{the data.frame from which the attributes should be copied.}

\item{target}{the target data.frame to which the attributes should be copied.}

\item{whichVars}{Optional: character vector of variables which attributes should
be copied. If \code{NULL}, all common variables will be used.}

\item{whichAttrs}{Optional: character vector of attributes which should be copied.
If \code{NULL}, all attributes from origin data.frame variables will be used.}
}
\value{
The target data.frame with additional attributes
}
\description{
Function allows to copy the attributes of (a subset of) common variables
from an origin data.frame to a target data.frame.
}
\examples{
mtcars2 <- mtcars
attr(mtcars[,"cyl"], "varLabel") <- "Number of cylinders"
mtcars2 <- transferAttributes(mtcars, mtcars2, whichVars = "cyl", whichAttrs="varLabel")

}
