% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unzip.R
\name{unzip.wa}
\alias{unzip.wa}
\title{Unzip files into desired directory}
\usage{
unzip.wa(
  zipfile,
  targetdir = tempdir(),
  unzip.exe = "c:/Program Files/tools/unzip.exe"
)
}
\arguments{
\item{zipfile}{Folder of the zip file which should be uncompressed.}

\item{targetdir}{The folder in which the content should be written.}

\item{unzip.exe}{The folder of the unzip.exe program.}
}
\value{
A character string (snippet) useful for formula generation.
}
\description{
This is a workaround function for \code{unzip} which provides
unexpected errors especially with data bases. \code{unzip.wa} calls the
\code{unzip.exe} via a temporary batch file.
}
\details{
Function uses the unzip.exe which is freely available at
\url{http://stahlworks.com/dev/index.php?tool=zipunzip#zipexamp}
}
\examples{
\dontrun{
### gives error
zip::unzip(zipfile = "q:/BT2016/BT/05_Anschreiben/03_Anschreiben_IQB_an_Eltern_Lehrer_Schulleiter/00_Archiv/20151110_finale_Unterlagen_Genehmigungsverfahren/LV_2016_Unterlagen_BB.zip", exdir=tempdir())
unzip.wa(zipfile = "q:/BT2016/BT/05_Anschreiben/03_Anschreiben_IQB_an_Eltern_Lehrer_Schulleiter/00_Archiv/20151110_finale_Unterlagen_Genehmigungsverfahren/LV_2016_Unterlagen_BB.zip")
}

}
