% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_xlsx.R
\name{update_xlsx}
\alias{update_xlsx}
\title{Update an \code{.xlsx} file.}
\usage{
update_xlsx(newDat, filePath, sheetName, idCol)
}
\arguments{
\item{newDat}{The new \code{data.frame}.}

\item{filePath}{The path to the \code{.xlsx} file..}

\item{sheetName}{The name of the sheet in the \code{.xlsx} file.}

\item{idCol}{Name of the unique identifier column.}
}
\value{
A list or a \code{data.frame}.
}
\description{
Update the content of an existing \code{.xlsx} file adding and removing rows in accordance to an identifier column. The content
of the columns of variables not in \code{idCol} is dropped.
}
\details{
This function is useful if, for example, a recode table or decision table is written to \code{.xlsx} but the underlying data structure changes.
In such cases it might be useful to update the \code{.xlsx} table, this means removing no longer existing variables and adding rows for
new variables. Use cases might be tables created by \code{eatGADS::getChangeMeta()} or \code{eatFDZ::reverse_check_docu()}.
}
\examples{
# tbd

}
