% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLabel.R
\name{convertLabel}
\alias{convertLabel}
\title{Transform SPSS variable and value labels into old ZKD convention.}
\usage{
convertLabel(spssList, stringsAsFactors = TRUE, useZkdConvention = TRUE)
}
\arguments{
\item{spssList}{An object created by \code{\link[foreign]{read.spss}}. Important: Using \code{\link[foreign]{read.spss}}, \code{to.data.frame} has to be FALSE.}

\item{stringsAsFactors}{Transform character variables into factors?}

\item{useZkdConvention}{Logical: Use ZKD convention, i.e. variable.labels are named 'varLabel', value.labels are named 'valLabel'.}
}
\value{
A \code{data.frame}.
}
\description{
By default, 'read.spss' from the foreign package uses variable.labels as attributes of the whole data.frame, value.labels as attribute of each specific
variable in the data.frame. \code{convertLabel} provides variable and value labels as variable attributes.
}
\examples{
file <- system.file("extdata", "Klauer.sav", package = "eatAnalysis")
dat  <- foreign::read.spss(file, to.data.frame=FALSE, use.value.labels = FALSE)
dat  <- convertLabel(dat)
str(dat)
}
\author{
Sebastian Weirich
}
