% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intGen.R
\name{intGen}
\alias{intGen}
\title{Generate interaction terms for lm formulas}
\usage{
intGen(vars, upto = 3)
}
\arguments{
\item{vars}{Character vector of variables}

\item{upto}{The maximum way of interactions which should be created.}
}
\value{
A character string (snippet) useful for formula generation.
}
\description{
Specify two-way or three-way interactions from a set of independent variables.
}
\examples{
vars <- paste0("Var_", letters[1:5])
### only two way interactions
frml <- intGen(vars = vars, upto = 2)

}
\author{
Sebastian Weirich
}
