% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{prop_table}
\alias{prop_table}
\title{Create a standard relative frequency table.}
\usage{
prop_table(vec, useNA = "no", round_perc = 1)
}
\arguments{
\item{vec}{A vector.}

\item{useNA}{How should missing values be treated in the table? Possible values: \code{no}, \code{ifany}, \code{always}.}

\item{round_perc}{With how many decimals should the percantages be given?}
}
\value{
Returns a data.frame with 1 row.
}
\description{
Create a simple relative frequency table with percentages.
}
\details{
Function is meant to be able to easily \code{rbind} various tables.
}
\examples{
# Example data set
#to be done

}
