% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2file.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Save to xlsx with multiple sheets.}
\usage{
write_xlsx(df_list, filePath, row.names = TRUE, col.names = TRUE)
}
\arguments{
\item{df_list}{A named list with data.frames to be written to xlsx. The names of the list become the sheet names.}

\item{filePath}{The path to the output file.}

\item{row.names}{A logical value indicating whether the row names of x are to be written along with x to the file.}

\item{col.names}{A logical value indicating whether the column names of x are to be written along with x to the file.}
}
\description{
Create an xlsx file with multiple sheets.
}
\details{
An existing file is overwritten
}
\examples{
# Example data set
#to be done

}
